+++ CHANGES.md
@@ -34,6 +34,7 @@ Release Notes.
 - Add lifecycle management for the node.
 - Property: Introduce the schema style to the property.
 - Add time range parameters to stream index filter.
+- UI: Add the `stages` to groups.
 - Add time range return value from stream local index filter.
 - Deduplicate the documents on building the series index.